move to go modules
It's been a long while since working on this project. In the meantime,
go modules are now mainstream. Move to using go modules and internal
packages rather than vendoring. This is a standalone program, we do not
expect any packages to be imported.
