+++ .travis.yml
@@ -1,8 +1,7 @@
 language: go
 
 go:
-  - "1.10.x"
+  - "1.16.x"
 
 after_success:
   - bash <(curl -s https://codecov.io/bash)
-
