+++ go.mod
@@ -1,0 +1,11 @@
+module github.com/fitzy101/streammon
+
+go 1.16
+
+require (
+	github.com/fsnotify/fsnotify v1.5.1 // indirect
+	github.com/hpcloud/tail v1.0.0
+	golang.org/x/sys v0.0.0-20220114195835-da31bd327af9 // indirect
+	gopkg.in/fsnotify.v1 v1.4.7 // indirect
+	gopkg.in/tomb.v1 v1.0.0-20141024135613-dd632973f1e7 // indirect
+)
