+++ internal/stream/rw.go
