+++ internal/stream/rw_test.go
