+++ internal/stream/stream.go
