+++ internal/stream/stream_test.go
